Update 1_makecsv.py
This revision modifies the Python script to enable the CSV file-making even though the `._*` files are both existed in the target folder. The script now excludes these files from the list of files to be processed. This update ensures that the CSV file is generated correctly even if there are hidden files in the directory.